Categoria apenas é elimininada por quem a criou

+++ app/src/main/java/pt/isec/amov/ui/screens/ManageCategory.kt
@@ -137,10 +137,12 @@ fun ManageCategoryScreen(navController: NavHostController, vm: ActionsViewModel,
                                 )
                             }
 
+                            if (vm.user.value?.email == it.createdBy){
+                                DeleteDialog(onClick = {
+                                    vm.deleteCategory(it.name);
+                                })
+                            }
 
-                                    DeleteDialog(onClick = {
-                                        vm.deleteCategory(it.name);
-                                    })
